Changed SOIL2 to regular stb image
Also fixed ChangeTexture and removed fullscreen (for now)

+++ Tin/include/Tin/Renderer/Renderer.hpp
@@ -19,7 +19,7 @@ namespace Tin {
         // Clears the color, depth and stencil buffers
         void Clear() const;
         // Captures an area and writes it to an image
-        void SaveScreenshot(const std::string& filename, const glm::vec2& position, const glm::vec2& size) const;
+        void SaveScreenshot(const std::string& filename, const glm::ivec2& position, const glm::ivec2& size) const;
 
         // Sets the opengl viewport size
         void SetViewportSize(const glm::vec2& size) const;
